+++ scripts/install_scikits_odes.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-SUNDIALS_URL=https://github.com/LLNL/sundials/archive/v4.1.0.tar.gz
-SUNDIALS_NAME=sundials-4.1.0.tar.gz
+SUNDIALS_URL=https://github.com/LLNL/sundials/archive/v5.0.0.tar.gz
+SUNDIALS_NAME=sundials-5.0.0.tar.gz
 CURRENT_DIR=`pwd`
 TMP_DIR=$CURRENT_DIR/tmp
 mkdir $TMP_DIR
@@ -10,9 +10,9 @@ INSTALL_DIR=$CURRENT_DIR/sundials
 cd $TMP_DIR
 wget $SUNDIALS_URL -O $SUNDIALS_NAME
 tar -xvf $SUNDIALS_NAME
-mkdir build-sundials-4.1.0
-cd build-sundials-4.1.0/
-cmake -DLAPACK_ENABLE=ON -DSUNDIALS_INDEX_TYPE=int32_t -DBUILD_ARKODE:BOOL=OFF -DEXAMPLES_ENABLE:BOOL=OFF -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR ../sundials-4.1.0/
+mkdir build-sundials-5.0.0
+cd build-sundials-5.0.0/
+cmake -DLAPACK_ENABLE=ON -DSUNDIALS_INDEX_TYPE=int32_t -DBUILD_ARKODE:BOOL=OFF -DEXAMPLES_ENABLE:BOOL=OFF -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR ../sundials-5.0.0/
 make install
 cd $CURRENT_DIR
 rm -rf $TMP_DIR